requirements.txt
dnspython==1.16.0

PyNaCl==1.3.0

async-timeout==3.0.1

discord.py==1.4.1

tmdbv3api

rarbgapi

python-dotenv

regex

aiohttp

rotten_tomatoes_client

chess.com
